remove api dependency from common
